Cocoa modifiers are NSUIntegers

+++ Classes/NDKeyboardLayout.h
@@ -43,20 +43,20 @@ extern NSString		* const NDKeyboardLayoutPreviousKeyboardLayoutUserInfoKey;
  Converts modifiers to their unicode characters
  @param modifierFlags ￼one or more of the flags <tt>NSShiftKeyMask</tt>, <tt>NSControlKeyMask</tt>, <tt>NSAlternateKeyMask</tt>, <tt>NSCommandKeyMask</tt>
  */
-NSString *NDStringForModifiers(UInt32 modifierFlags);
+NSString *NDStringForModifiers(NSUInteger modifierFlags);
 
 /*!
 	@function NDCocoaModifierFlagsForCarbonModifierFlags
 	Convert Carbon modifer flags to Cocoa modifier flags.
 	@param modifierFlags one or more of the flags <tt>shiftKey</tt>, <tt>controlKey</tt>, <tt>optionKey</tt>, <tt>cmdKey</tt>
  */
-NSUInteger NDCocoaModifierFlagsForCarbonModifierFlags( NSUInteger modifierFlags );
+NSUInteger NDCocoaModifierFlagsForCarbonModifierFlags( UInt32 modifierFlags );
 /*!
 	@function NDCarbonModifierFlagsForCocoaModifierFlags
 	Convert Cocoa modifer flags to Carbon modifier flags.
 	@param modifierFlags ￼one or more of the flags <tt>NSShiftKeyMask</tt>, <tt>NSControlKeyMask</tt>, <tt>NSAlternateKeyMask</tt>, <tt>NSCommandKeyMask</tt>
  */
-NSUInteger NDCarbonModifierFlagsForCocoaModifierFlags( NSUInteger modifierFlags );
+UInt32 NDCarbonModifierFlagsForCocoaModifierFlags( NSUInteger modifierFlags );
 
 /*!
 	@class NDKeyboardLayout
@@ -105,7 +105,7 @@ NSUInteger NDCarbonModifierFlagsForCocoaModifierFlags( NSUInteger modifierFlags 
 	@param character The unmodified character on the keyboard.
 	@param modifierFlags Modifier flags <tt>NSControlKeyMask</tt>, <tt>NSAlternateKeyMask</tt>, <tt>NSShiftKeyMask</tt>, <tt>NSCommandKeyMask</tt> and <tt>NSNumericPadKeyMask</tt>.
  */
-- (NSString*)stringForCharacter:(unichar)character modifierFlags:(UInt32)modifierFlags;
+- (NSString*)stringForCharacter:(unichar)character modifierFlags:(NSUInteger)modifierFlags;
 /*!
 	@method stringForKeyCode:modifierFlags:
 	@abstract Get a string for display purposes. 
@@ -113,7 +113,7 @@ NSUInteger NDCarbonModifierFlagsForCocoaModifierFlags( NSUInteger modifierFlags 
 	@param keyCode A value specifying the virtual key code that is to be translated. For ADB keyboards, virtual key codes are in the range from 0 to 127.
 	@param modifierFlags Modifier flags <tt>NSControlKeyMask</tt>, <tt>NSAlternateKeyMask</tt>, <tt>NSShiftKeyMask</tt>, <tt>NSCommandKeyMask</tt> and <tt>NSNumericPadKeyMask</tt>.
  */
-- (NSString*)stringForKeyCode:(UInt16)keyCode modifierFlags:(UInt32)modifierFlags;
+- (NSString*)stringForKeyCode:(UInt16)keyCode modifierFlags:(NSUInteger)modifierFlags;
 /*!
 	@method characterForKeyCode:
 	@abstract Get the key character for a given key code.
